Add compiledEntry field to Pattern structure
Introduce a field associated with the Pattern data structure for
holding the entry point to a (natively) compiled version of the
pattern. JIT compilers can write the entry point of the compiled
pattern to this field, and the virtual machine can be modified
to invoke the entry point if a non-zero value is present. By
default, the field is always initialized to zero.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>