Add makefile targets to clone omr and build jitbuilder
Added two makefile targets to clone omr as a peer to the
rosie-lpeg directory. After cloning, it configures and
then builds the jitbuilder library inside the cloned
omr repository. Because the configure command is
different for each platform, there are omr_macosx and
omr_linux targets. Once this step is complete, there
is an omr/jitbuilder/release/libjitbuilder.a library
file that can be used by subsequent build steps.

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>