Implement lpeg JIT compiler
This is a big commit. The primary change is the introduction
of the lpjit.cpp and lpjit.hpp files, which implement a JitBuilder
based JIT compiler for lpeg. It currently handles all but two
bytecodes (CALL and RET). These two bytecodes aren't used by
the Rosie pattern language. This JIT required a few connection
points with code in lpvm.c. Specifically, several helper functions
are called by the generated code, and one bytecode handler
was also moved into a function to make it easy for the JITted
code to call it.

The JIT compiler itself is driven by the function buildIL() in
the LpMatcher class. This function creates a BytecodeBuilder
object for each bytecode in the method and then uses the
JitBuilder worklist service to visit each bytecode once to
generate native code for that bytecode. Each bytecode handler
is implemented by a member function of the LpMatcher class
called do<BytecodeName (e.g. doTestChar, doSet). Compiled
methods have the same interface as the match() function that
implements the bytecode interpreter, which makes it very easy
to perform a patten match using the compiled code rather than
the interpreter.

Lua entry points were added to the lpeg library to be able to compile
patterns ("compile" implemented by lp_compile() ) and to do matches
using the compiled code for patterns ("compileAndMatch" implemented
by lp_compileAndMatch(). The pre-existing "match" function is
modified to check for a compiled entry point and to call the compiled
code rather than intepreting the pattern if such an entry point
exists. Also added a "cleanup" function that can be called to
shut the JIT down (among other things).

I added clock_gettime() calls around the different match implementations
so that the cumulative match time can be measured. This facility
is off by default, but can be activated by defining MEASURE in
lpvm.c .

Finally, a lua test file called test-lpeg.lua (similar to test.lua)
will use the JIT compiler to match patterns when the cs2str() function
is called. There are a number of pattern matching tests that will
use the JIT compiler. Since these tests do not tend to reuse the
same pattern, JIT compiling them does not help performance (in fact,
it slows the test down) but it does ensure that the JIT compiler
is exercised and tested.

Signed-off- by: Mark Stoodley <mstoodle@ca.ibm.com>